Hash Tables: Ransom Note
Harold is a kidnapper who wrote a ransom note, but now he is worried it
will be traced back to him through his handwriting. He found a magazine
and wants to know if he can cut out whole words from it and use them to
create an untraceable replica of his ransom note. The words in his note
are case-sensitive and he must use only whole words available in the
magazine. He cannot use substrings or concatenation to create the words
he needs.Harold is a kidnapper who wrote a ransom note, but now he is
worried it will be traced back to him through his handwriting. He found
a magazine and wants to know if he can cut out whole words from it and
use them to create an untraceable replica of his ransom note. The words
in his note are case-sensitive and he must use only whole words
available in the magazine. He cannot use substrings or concatenation to
create the words he needs.